% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{filter_chain}
\alias{filter_chain}
\title{Filter MCMC Chain (Burn-in and Thinning)}
\usage{
filter_chain(sample_list, burn_in, iters, thin)
}
\arguments{
\item{sample_list}{List of MCMC samples.}

\item{burn_in}{Integer. Burn-in iterations.}

\item{iters}{Integer. Total iterations.}

\item{thin}{Integer. Thinning interval.}
}
\value{
Filtered sample list.
}
\description{
Internal: Remove burn-in and apply thinning to MCMC sample list.
}
\keyword{internal}
