% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gen_dummy}
\alias{gen_dummy}
\title{Generate Dummy (Indicator) Matrix}
\usage{
gen_dummy(x, n_cat, intercept = FALSE)
}
\arguments{
\item{x}{Integer vector of categories (1-based).}

\item{n_cat}{Integer. Number of categories.}

\item{intercept}{Logical. Include intercept column? Default is FALSE.}
}
\value{
Numeric matrix of dummy variables.
}
\description{
Internal: Create a dummy variable (indicator) matrix for a categorical variable, with optional intercept.
}
