% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbfseq-package.R
\docType{package}
\name{mbfseq}
\alias{mbfseq-package}
\alias{mbfseq}
\title{mbfseq package}
\description{
Description of your package here.
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
